happy path and basic negative test cases created and passed for ItemController

+++ src/main/java/com/mdsujan/restPostgres/service/ItemService.java
@@ -205,7 +205,7 @@ public class ItemService {
 
     public PaginatedResponse getAllItemsPaginated(Integer pageSize, Integer pageNum) {
         List<Item> items = itemRepository.findAll(PageRequest.of(pageNum - 1, pageSize)).getContent();
-        Long maxPages = itemRepository.count() / pageSize;
+        Integer maxPages = Math.toIntExact(itemRepository.count() / pageSize);
 
         return new PaginatedResponse(items, maxPages, pageNum, pageSize);
     }
